requirements.txt
fastapi
uvicorn
line-bot-sdk
selenium
webdriver-manager
lxml  